Switch HTTP client to http.rb
Primarily this is because we are soon going to depend on the GeoIP2-ruby
gem, which uses http.rb already, and it seems undesirable to have to
pull in two different HTTP clients. This also has the benefit for
maintainers that we only need to work with a single HTTP client across
both gems. The implementation ends up being simpler as well.

This is technically a breaking change but I am doing it in such a way as
to hopefully have no actual breaking change on users, as the breaking
changes are not in spots likely to be used externally.