Depend on maxmind-geoip2
We use classes from this gem as part of our responses.

We wanted to do this previously but could not as the maxmind-geoip2
depended on a newer version of Ruby than we were supporting. Now that is
no longer an issue. This allows us to remove the essentially vendored
files we had from the gem as a workaround.